[chore] :  refactor package structure
