+++ src/main/java/nextstep/courses/domain/image/Image.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.image;
 
 import java.util.Objects;
