+++ src/main/java/nextstep/courses/domain/session/FreeSession.java
@@ -1,4 +1,7 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
+
+import nextstep.courses.domain.image.Image;
+import nextstep.courses.domain.student.Student;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -34,7 +37,7 @@ public class FreeSession extends Session {
     }
 
     @Override
-    protected void register(Registration registration) {
+    public void register(Registration registration) {
         if (!isAvailableForRegistration()) {
             throw new IllegalStateException("Can't register session");
         }
