+++ src/main/java/nextstep/courses/domain/session/SessionType.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import java.util.Arrays;
