+++ src/test/java/nextstep/courses/domain/SessionCapacityTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.session.SessionCapacity;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -21,12 +22,12 @@ public class SessionCapacityTest {
         int capacity = 0;
 
         Assertions.assertThatThrownBy(() -> new SessionCapacity(capacity))
-            .isInstanceOf(IllegalArgumentException.class);
+                .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     @DisplayName("수강행 추가")
-    void addCurrentCountTest(){
+    void addCurrentCountTest() {
         SessionCapacity sessionCapacity = new SessionCapacity(10);
         sessionCapacity.increase();
 
@@ -36,12 +37,12 @@ public class SessionCapacityTest {
 
     @Test
     @DisplayName("수강생 초과 체크")
-    void checkOverCurrentCountTest(){
+    void checkOverCurrentCountTest() {
         SessionCapacity sessionCapacity = new SessionCapacity(1);
 
         sessionCapacity.increase();
 
         Assertions.assertThatThrownBy(sessionCapacity::increase)
-            .isInstanceOf(IllegalArgumentException.class);
+                .isInstanceOf(IllegalArgumentException.class);
     }
 }
